+++ examples/js/hacker-news/client/containers/App.js
@@ -67,7 +67,7 @@ const App = ({ route }) => (
   <div>
     <Header
       title="reSolve Hacker News"
-      favicon="/favicon.ico"
+      favicon="/favicon.png"
       css={['/style.css']}
     />
     <ContentRoot>