+++ examples/ts/shopping-list-redux/client/ssr.tsx
@@ -55,7 +55,7 @@ const ssrHandler = async (serverContext, req, res) => {
 
     const initialState = store.getState()
     const bundleUrl = getStaticBasedPath(rootPath, staticPath, 'index.js')
-    const faviconUrl = getStaticBasedPath(rootPath, staticPath, 'favicon.ico')
+    const faviconUrl = getStaticBasedPath(rootPath, staticPath, 'favicon.png')
 
     const helmet = Helmet.renderStatic()