+++ packages/tools/testing-tools/src/flow/read-model/make-test-environment.ts
@@ -128,7 +128,7 @@ export const makeTestEnvironment = (
           ADAPTER_NAME: actualAdapter,
         },
         getVacantTimeInMillis: () => 0x7fffffff,
-        invokeEventSubscriberAsync: async () => {
+        invokeBuildAsync: async () => {
           isNext = true
         },
         eventstoreAdapter,