+++ templates/js/styled-components/client/components/App.js
@@ -6,7 +6,7 @@ const App = ({ route }) => (
     <Header
       title="reSolve Styled-Components Example"
       name="Styled-Components Example"
-      favicon="/favicon.ico"
+      favicon="/favicon.png"
       css={['/bootstrap.min.css']}
     />
     {renderRoutes(route.routes)}