+++ examples/ts/hacker-news/import/index.ts
@@ -6,7 +6,7 @@ import path from 'path'
 
 const runImport = (importConfig) => {
   const port = importConfig.hasOwnProperty('port')
-    ? String(importConfig.port)
+    ? String(importConfig.port.defaultValue)
     : process.env.PORT || '3000'
 
   const appUrl = `http://localhost:${path.join(port, importConfig.rootPath)}`