+++ packages/runtime/runtime/src/common/init-resolve.js
@@ -21,7 +21,7 @@ const initResolve = async (resolve) => {
   } = resolve.assemblies
 
   const {
-    invokeEventSubscriberAsync,
+    invokeBuildAsync,
     applicationName,
     uploader,
     scheduler,
@@ -93,8 +93,8 @@ const initResolve = async (resolve) => {
     secretsManager,
     eventstore: eventstoreAdapter,
     hooks: {
-      postSaveEvent: async (aggregate, command, event) => {
-        await onCommandExecuted(event, command)
+      postSaveEvent: async (aggregate, command, event, eventWithCursor) => {
+        await onCommandExecuted(event, command, eventWithCursor)
         return false
       },
     },
@@ -116,7 +116,7 @@ const initResolve = async (resolve) => {
   })
 
   const executeQuery = createQueryExecutor({
-    invokeEventSubscriberAsync,
+    invokeBuildAsync,
     applicationName,
     eventstoreAdapter,
     readModelConnectors,
@@ -138,7 +138,7 @@ const initResolve = async (resolve) => {
   })
 
   const executeSaga = createSagaExecutor({
-    invokeEventSubscriberAsync,
+    invokeBuildAsync,
     applicationName,
     executeCommand,
     executeQuery,