+++ templates/ts/postcss/client/ssr.tsx
@@ -34,7 +34,7 @@ const ssrHandler = async (serverContext, req, res) => {
     )
 
     const bundleUrl = getStaticBasedPath(rootPath, staticPath, 'index.js')
-    const faviconUrl = getStaticBasedPath(rootPath, staticPath, 'favicon.ico')
+    const faviconUrl = getStaticBasedPath(rootPath, staticPath, 'favicon.png')
     const helmet = Helmet.renderStatic()
 
     const markupHtml =