+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Upcoming](https://github.com/reimagined/resolve/tree/HEAD)
+
+[Full Changelog](https://github.com/reimagined/resolve/compare/V0.30.5...HEAD)
+
+**New features:**
+
+- Upgrade examples [\#1902](https://github.com/reimagined/resolve/pull/1902)
+
 ## [V0.30.5](https://github.com/reimagined/resolve/tree/V0.30.5) (2021-06-08)
 
 [Full Changelog](https://github.com/reimagined/resolve/compare/V0.30.4...V0.30.5)