+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Upgrade examples [\#1902](https://github.com/reimagined/resolve/pull/1902)
 
+**Bug fixes:**
+
+- The loadEvents function does not work if 'startTime' and 'endTime' parameters are specified [\#1903](https://github.com/reimagined/resolve/issues/1903)
+
 ## [V0.30.5](https://github.com/reimagined/resolve/tree/V0.30.5) (2021-06-08)
 
 [Full Changelog](https://github.com/reimagined/resolve/compare/V0.30.4...V0.30.5)